chore: rename Enterprise reporter to DashboardReporter
